requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
python-jose
python-multipart
pytest
pytest-cov
httpx
python-dotenv
locust